requirements.txt
faker
fastapi
uvicorn
fastapi_pagination
pytest
pyyaml
requests
pyspark
delta-spark